pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "search-relevance"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "nltk",
    "matplotlib",
    "pyyaml",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]